move BoundType into own file & fix failing tests
